+++ src/config/plain.c
@@ -1590,7 +1590,7 @@ char *build_json_from_tokens(config_parser_stat *wstokens, uint64_t token_count)
 									continue;
 								}
 
-								if (!strcmp(operator_name, "field") || !strcmp(operator_name, "jpath") || !strcmp(operator_name, "valid_status_codes") || !strcmp(operator_name, "servers") || !strcmp(operator_name, "sharding_key") || !strcmp(operator_name, "match"))
+								if (!strcmp(operator_name, "field") || !strcmp(operator_name, "jpath") || !strcmp(operator_name, "except") || !strcmp(operator_name, "valid_status_codes") || !strcmp(operator_name, "servers") || !strcmp(operator_name, "sharding_key") || !strcmp(operator_name, "match"))
 								{
 									json_t *arg_json = json_object_get(operator_json, operator_name);
 									if (!arg_json)