+++ src/query/type.h
@@ -1,6 +1,7 @@
 #pragma once
 #include "dstructures/tommy.h"
 #include "events/context_arg.h"
+#include "common/selector.h"
 #include <jansson.h>
 typedef struct query_field
 {
@@ -29,6 +30,7 @@ typedef struct query_node
 	char *datasource;
 	context_arg *carg;
 	alligator_ht *labels;
+	match_rules *except;
 
 	tommy_node node;
 } query_node;
@@ -44,6 +46,8 @@ query_ds* query_get(char *datasource);
 query_field* query_field_get(alligator_ht *qf_hash, char *key);
 alligator_ht* query_get_field(json_t *jfield);
 query_node *query_get_node(query_ds *qds, char *make);
+int query_except_match(query_node *qn, const char *name);
+int query_ds_except_db(query_ds *qds, const char *dbname);
 void query_node_del(query_node *qn);
 void query_set_values(query_node *qn);
 int query_del(json_t *query);